Reduce offset
Reduce offset from 82_104_121_68_114_000 to 82_68_114_000

+++ RhythmDoctor.Archipelago/World/Structures/Item.cs
@@ -6,7 +6,7 @@ public struct Item
   public string Name;
 
   [YamlMember(Alias = "id")]
-  public ulong ID;
+  public uint ID;
 
   [YamlMember(Alias = "classification")]
   public ItemClassification Classification;
